finition 3a

+++ scripts/3a-save.py
@@ -15,38 +15,53 @@ import os
 
 ##### VARIABLES #####
 
-path_my_directory = os.path.expanduser('~/B2-Python')
-path_data = os.path.expanduser('~/data')
-path_my_backup = os.path.expanduser('~/backup_B2-Python')
+path_my_directory = '/root/B2-Python'
+path_data = '/root/data'
+path_my_backup = '/tmp/backup_B2-Python'
+
+msgErrorPerm = 'Vous n avez pas la permission !\n'
+msgSaved = 'Sauvegarde effectuer !\n'
+msgSavedExisting = 'Votre sauvegarde existe deja !\n'
 
 ##### FUNCTIONS #####
 
 def youcant(sig, frame):
-	goodbye()
+	if os.path.isfile(path_my_backup):
+		os.remove(path_my_backup)
+	sys.stdout('Sauvegarde annuler !\n')
 	sys.exit(0)
 
 def makeArchive(src_dir, archive_dir):
-	shutil.make_archive(archive_dir, 'gztar', src_dir)
+	if os.access(src_dir, os.R_OK):
+		shutil.make_archive(archive_dir, 'gztar', src_dir)
+	else:
+		sys.stderr.write(msgErrorPerm)
 
 ##### SCRIPT #####
 
 signal.signal(signal.SIGINT, youcant)
 
-if not os.path.exists(path_data):
-	os.mkdir(path_data)
+if os.access('/root', os.R_OK and os.W_OK):
 
-makeArchive(path_my_directory, path_my_backup)
-path_my_backup += '.tar.gz'
-#path_my_backup = os.path.expanduser('~/data/backup_B2-Python'+'.tar.gz')
-if os.path.isfile(path_data + '/backup_B2-Python.tar.gz'):
-	new_file = gzip.open(path_my_backup)
-	old_file = gzip.open(path_data+'/backup_B2-Python.tar.gz')
-	if new_file.read() == old_file.read():
-		sys.stdout.write('Existe deja !\n')
-		os.remove(path_my_backup)
+	if not os.path.isdir(path_data):
+		os.mkdir(path_data)
+
+	makeArchive(path_my_directory, path_my_backup)
+	path_my_backup += '.tar.gz'
+
+	if os.path.isfile(path_data + '/backup_B2-Python.tar.gz'):
+		new_file = gzip.open(path_my_backup)
+		old_file = gzip.open(path_data+'/backup_B2-Python.tar.gz')
+		if new_file.read() == old_file.read():
+			sys.stdout.write(msgSavedExisting)
+			os.remove(path_my_backup)
+		else:
+			os.remove(path_data+'/backup_B2-Python.tar.gz')
+			shutil.move(path_my_backup, path_data)
+			sys.stdout.write(msgSaved)
 	else:
-		sys.stdout.write('Sauvegarder \n')
-		os.remove(path_data+'/backup_B2-Python.tar.gz')
 		shutil.move(path_my_backup, path_data)
+		sys.stout.write(msgSaved)
+
 else:
-	shutil.move(path_my_backup, path_data)
+	sys.stderr.write(msgErrorPerm)
